+++ src/main/java/roomescape/controller/LoginController.java
@@ -7,9 +7,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.CheckResponseDto;
-import roomescape.dto.TokenRequest;
-import roomescape.dto.TokenResponse;
+import roomescape.dto.request.TokenRequest;
+import roomescape.dto.response.CheckResponseDto;
+import roomescape.dto.response.TokenResponse;
 import roomescape.service.AuthService;
 import roomescape.token.Cookies;
