+++ src/main/java/roomescape/controller/MemberController.java
@@ -5,8 +5,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.MemberRequestDto;
-import roomescape.dto.MemberResponseDto;
+import roomescape.dto.request.MemberRequestDto;
+import roomescape.dto.response.MemberResponseDto;
 import roomescape.service.MemberService;
 
 @RestController
