+++ src/main/java/roomescape/dao/ReservationTimeDao.java
@@ -3,7 +3,7 @@ package roomescape.dao;
 import java.util.List;
 import java.util.Optional;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.BookedReservationTimeResponseDto;
+import roomescape.dto.response.BookedReservationTimeResponseDto;
 
 public interface ReservationTimeDao {
