+++ src/main/java/roomescape/dto/request/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 public record MemberRequestDto(String name, String email, String password) {
