+++ src/main/java/roomescape/dto/request/ReservationTimeRequestDto.java
@@ -1,6 +1,5 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
-import java.time.LocalDate;
 import java.time.LocalTime;
 import java.time.format.DateTimeFormatter;
 import java.time.format.DateTimeParseException;
