+++ src/main/java/roomescape/dto/request/ThemeRequestDto.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.request;
 
 public record ThemeRequestDto(String name, String description, String thumbnail) {
