+++ src/main/java/roomescape/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.response;
 
 public record TokenResponse(String token) {
