+++ src/test/java/roomescape/dto/ReservationTimeRequestDtoTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import roomescape.dto.request.ReservationTimeRequestDto;
 
 class ReservationTimeRequestDtoTest {
